move holocene helpers to separate file
