+++ op-e2e/actions/proofs/holocene_frame_test.go
@@ -7,44 +7,9 @@ import (
 	actionsHelpers "github.com/ethereum-optimism/optimism/op-e2e/actions/helpers"
 	"github.com/ethereum-optimism/optimism/op-e2e/actions/proofs/helpers"
 	"github.com/ethereum-optimism/optimism/op-program/client/claim"
-	"github.com/ethereum-optimism/optimism/op-service/eth"
-	"github.com/ethereum-optimism/optimism/op-service/testlog"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
 )
-
-type logExpectations struct {
-	role   string
-	filter string
-	num    int
-}
-type expectations struct {
-	safeHead uint64
-	logs     []logExpectations
-}
-type holoceneExpectations struct {
-	preHolocene, holocene expectations
-}
-
-func (h holoceneExpectations) RequireExpectedProgressAndLogs(t actionsHelpers.StatefulTesting, actualSafeHead eth.L2BlockRef, isHolocene bool, engine *actionsHelpers.L2Engine, logs *testlog.CapturingHandler) {
-	var exp expectations
-	if isHolocene {
-		exp = h.holocene
-	} else {
-		exp = h.preHolocene
-	}
-
-	require.Equal(t, exp.safeHead, actualSafeHead.Number)
-	expectedHash := engine.L2Chain().GetBlockByNumber(exp.safeHead).Hash()
-	require.Equal(t, expectedHash, actualSafeHead.Hash)
-
-	for _, l := range exp.logs {
-		t.Helper()
-		recs := logs.FindLogs(testlog.NewMessageContainsFilter(l.filter), testlog.NewAttributesFilter("role", l.role))
-		require.Len(t, recs, l.num, "searching for %d instances of '%s' in logs from role %s", l.num, l.filter, l.role)
-	}
-
-}
 
 func Test_ProgramAction_HoloceneFrames(gt *testing.T) {
 	type testCase struct {
@@ -71,7 +36,8 @@ func Test_ProgramAction_HoloceneFrames(gt *testing.T) {
 			name: "disordered-a", frames: []uint{2, 1, 0},
 			holoceneExpectations: holoceneExpectations{
 				preHolocene: expectations{safeHead: 3}, // frames are buffered, so ordering does not matter
-				holocene:    expectations{safeHead: 0}, // non-first frames will be dropped b/c it is the first seen with that channel Id. The safe head won't move until the channel is closed/completed.
+				holocene: expectations{safeHead: 0, // non-first frames will be dropped b/c it is the first seen with that channel Id. The safe head won't move until the channel is closed/completed.
+					logs: sequencerOnce("dropping future batch")},
 			},
 		},
 		{
